refactor(activity-stepper): Renaming & moving step shift calculation to a separate class
https://mindlogger.atlassian.net/browse/M2-4234
