+++ src/features/pass-survey/model/hooks/useActivityStepper.ts
@@ -4,9 +4,13 @@ import {
   onIncorrectAnswerGiven,
 } from '../../lib';
 import AnswerValidator from '../AnswerValidator';
-import PipelineVisibilityChecker from '../PipelineVisibilityChecker';
+import StepperUtils from '../StepperUtils';
 
-const ItemsTakePartInCL: ActivityItemType[] = ['Radio', 'Checkbox', 'Slider'];
+const ConditionalLogicItems: ActivityItemType[] = [
+  'Radio',
+  'Checkbox',
+  'Slider',
+];
 
 function useActivityStepper(state: ActivityState | undefined) {
   const step = state?.step ?? 0;
@@ -44,13 +48,11 @@ function useActivityStepper(state: ActivityState | undefined) {
   const showBottomNavigation = !showTopNavigation;
   const showWatermark = !isSplashStep && !showTopNavigation;
 
-  const isItemAPartOfCL = ItemsTakePartInCL.includes(
+  const isConditionalLogicItem = ConditionalLogicItems.includes(
     currentPipelineItem!?.type,
   );
 
   const answerValidator = AnswerValidator(state);
-
-  const visibilityChecker = PipelineVisibilityChecker(items, answers);
 
   function isValid() {
     const valid = answerValidator.isCorrect();
@@ -62,35 +64,9 @@ function useActivityStepper(state: ActivityState | undefined) {
     return valid;
   }
 
-  function getNextStepShift(direction: 'forwards' | 'backwards') {
-    let shift = 1;
-
-    while (true) {
-      let nextStep = direction === 'forwards' ? step + shift : step - shift;
-
-      if (nextStep > items.length - 1) {
-        shift = items.length;
-        break;
-      }
-
-      if (nextStep < 0) {
-        break;
-      }
-
-      const isItemVisible = visibilityChecker.isItemVisible(nextStep);
-
-      if (isItemVisible) {
-        break;
-      } else {
-        shift++;
-      }
-    }
-
-    return shift;
-  }
-
   function getNextButtonText() {
-    const shift = getNextStepShift('forwards');
+    const stepperUtils = new StepperUtils(state!);
+    const shift = stepperUtils.getNextStepShift('forwards');
 
     if (shift >= items.length) {
       return 'activity_navigation:done';
@@ -107,7 +83,7 @@ function useActivityStepper(state: ActivityState | undefined) {
     isTutorialStep,
     isFirstStep,
     isLastStep,
-    isItemAPartOfCL,
+    isConditionalLogicItem,
 
     canSkip,
     canMoveNext,
